+++ lib/terrazine/compilers/base.rb
@@ -4,44 +4,7 @@ module Terrazine
   module Compilers
     class Base
       CONSTRUCTOR_CLASS = Constructor
-
-      class << self
-        def def_multi(*args, &method)
-          distinction = if args.count > 1
-                          method_name = args.first
-                          args.second
-                        else
-                          method_name = :multimethod
-                          args.first
-                        end
-          multimethod(method_name).add_method(distinction, &method)
-        end
-
-        def def_default_multi(method_name = :multimethod, &method)
-          multimethod(method_name).assign_default(&method)
-        end
-
-        def multimethod(method_name)
-          method = "@#{method_name}"
-          return instance_variable_get(method) if instance_variable_defined?(method)
-
-          initialize_multi(:class, method_name)
-        end
-
-        def initialize_multi(differ = :class, method_name = :multimethod)
-          m = instance_variable_set("@#{method_name}", Multimethods.new(differ))
-          define_method(method_name) do |*args|
-            method = m.fetch_method(args.first) || m.default_method
-            instance_exec(*args, &method)
-          end
-          define_method("#{method_name}_by") do |*args|
-            method = m.fetch_method(args.first)
-            next instance_exec(*args.drop(1), &method) if method
-            instance_exec(args.first, *args.drop(1), &m.default_method)
-          end
-          m
-        end
-      end
+      extend MultimethodsInterface
 
       def initialize(options)
         @options = options
