+++ spec/compilers/conditions_spec.rb
@@ -18,6 +18,14 @@ describe 'Compilers::Condition' do
       let(:result) { 'u.id = m.user_id' }
       it { is_expected.to eq result }
     end
+
+    context 'combined' do
+      let(:structure) do
+        [:_and, [:_eq, :u__id, :m__user_id], { m__id: :f__m_id }]
+      end
+      let(:result) { 'u.id = m.user_id AND m.id = f.m_id' }
+      it { is_expected.to eq result }
+    end
   end
 
   context 'Hash' do
@@ -33,12 +41,6 @@ describe 'Compilers::Condition' do
       it { is_expected.to eq result }
     end
 
-    # context 'as array of equalities?' do
-    #   let(:structure) { { u__id: [:m__user_id, :f__user_id] } }
-    #   let(:result) { 'u.id = m.user_id AND u.id = f.user_id' }
-    #   it { is_expected.to eq result }
-    # end
-
     context 'as in statement' do
       let(:structure) { { u__id: [:m__user_id, :f__user_id] } }
       let(:result) { 'u.id IN (m.user_id, f.user_id)' }
@@ -48,19 +50,7 @@ describe 'Compilers::Condition' do
 
   context 'String' do
     let(:structure) { 'true' }
-    let(:result) { structure }
+    let(:result) { structure.upcase }
     it { is_expected.to eq result }
   end
-
-  # context 'as sub query interpolation' do
-  #   let(:structure) { ['id IN ?', { select: :* }] }
-  #   let(:result) { 'id IN (SELECT * )' }
-  #   it { is_expected.to eq result }
-  # end
-
-  # context 'as params interpolation' do
-  #   let(:structure) { ['name = ?', 'Aeonax'] }
-  #   let(:result) { ['name = $1', 'Aeonax'] }
-  #   it { is_expected.to eq result }
-  # end
 end
